Added a new variable for program's event source identifier
Added a new global variable that will hold the program's event source identifier.  This will help to make sure that all events casted within the program are all unified, thus making it easy to not only subscribe to the events but to also unsubscribe as well.  This will allow the program to easily use powershell's events.