Logging object is now entirely Static
Greatly changed the entire class of 'Logging' to be entirely 'static'.  This change was necessary to allow the possibility of using events within the PowerShell engine.  Thus, other objects are now capable of using the member functions within Logging - without needing to create a new instance of the object.  The change to allow full static was necessary due to how the PowerShell engine works with custom events.