IOCommon is now static!
Greatly revised all member functions within IOCommon to be static functions.  With this change, other functions that are part of this software no longer need to create an instance of the IOCommon object in order to use the functions.  As such, functions that need a specific functionality within IOCommon - they can simply just call the desired function.  This should make using IOCommon's functions a bit easier.