All related objects append to IOCommon's latest changes
All objects that previously used IOCommon had to be revised to append to the latest changes of IOCommon.  Thus, with these changes, all related functions that previously created a new instance of the IOCommon object - no longer need to do so.  Instead, in order to use any functions from the IOCommon - just merely use them without needing an instance of the object before hand.