ProjectInformation is now entirely static
Because this class is not dynamic in anyway and only exists to provide the project's information (which is never changed), it is best to change this function as a static.  However, because it is now static, this class contains further changes.  This class no longer contains any Getter functions to retrieve the variable data, also the variables within this class are no longer considered 'hidden'.  Thus, the variables are now the only way to view the data directly.