All related objects append to ProjectInformation latest changes
Any object that used ProjectInformation no longer needs to create a new instance of the object, instead objects that require the information can merely read the data directly from the class itself.  With that, we can easily fetch the data without having to create a new instance of the object.