Added a new class: IOLoggingGateway
Added a new class named IOLoggingGateway.  This class is going to act as an gateway to both the Logging and the IOCommon classes, this will be helpful to minimize code duplication and mistakes of not using the required functions properly - such as logging and input and output and vice versa.  This class depends on IOCommon and Logging; without these objects - it will not object.

Added a new function DisplayMessage().  This function will take the message provided and allow it to be displayed and - once supported - logged.  This function is still working progress and still needs further developments.  This function is not yet finished as the logging class is not yet finished.