Added a new function DisplayMessage() (Short-Hand Ver)
Added a new function named DisplayMessage().  This function is merely a short-hand version of the DisplayMessage() which requires two arguments, the initial message and the message level.  This short-hand version merely only accepts the initial message without requiring the message level.
Because Powershell does not allow default values, at least the time of writing this, we will resort to overflowing the function.  Any messages accessing this short-hand version of the function will always have a message level as 'Standard'.  In order to use any other message levels, you must resort to using the initial full function named DisplayMessage(arg0, arg1).