Very WIP: Event Handling
WARNING: Not finished
Added a custom event that will execute when the source identifier is 'LogEvent-IOCommon'.  This event will allow functions within the class to trigger this event to quickly pass a message to centralize location specified during the event registeration.  However, the main implementation is still not finished.  The basic template works, but the idea of communicating to a specific function is not yet ready - more work and testing is required.