Added a new function CaptureLogEvent()
Added a new function named 'CaptureLogEvent()'.  This function will provide events to directly call this function to provide log information.  The message provided from the event will be passed to this function, which the message (mainly in a package form) will be transformed to one literal string.  The string, after it has been transformed, will be usable in WriteLogFile() in the Logging class.