Further refined fix in bbaf1f7
Further refined the fix in commit [ bbaf1f7f641dcb116118fed8dff8aa6a1075673d ]; if the directory or directories could not  be successfully created, then immediately prevent anything being written to a log file.  Previously, the fix merely ignored the status of the directory - just dismiss the state of the directory and try to write to a log file that may or may not exist.  This fix will take into account if the directory could not be created, maybe because of ACL or various other reasons, then immediately stop the function to avoid conflicts.