Fixed an issue with file paths containing spaces or special chars.
Fixed an issue that would cause 7Zip to use an incorrect file when the entire directory or file name contains a space or special characters.  To resolve this, just add quotes to the absolute or relative path.  With adding the quotes in the path string, 7Zip is able to distinguish the path specified and not additional commands or switches.