Overhauled EventLoggingRegister()
Renamed EventLoggingRegister() to RegisterEventLogging().  Previously this function was merely dedicated to the IOCommon object, it was not really available to other objects outside of the object itself.  Now with making this function more flexible for other objects to use freely, this can allow a much standardized version of registering the logging functionality to the Powershell engine.