Overhauled Destroy()
Greatly revised the actions performed in the Destroy() function that is part of the IOCommon object.  Instead of trying to unregister the custom event directly in the Destroy() function - while being hardcoded, the unregister event functionality has been shifted to its own function named 'UnregisterEven()'.  With this function, this will allow other objects to take advantage of this functionality and help make it a standard to this program.