Added a new function TriggerEvent()
Added a new function named TriggerEvent(), which will help to centralize the even trigger process.  In order for this to work, the source identifier (the first parameter) must already be registered to the PowerShell engine.  Additionally, this function only works for customized powershell events.  This function will help to minimize code redundancy and to centralize the functionality for all objects within the software.