Added a new function TriggerEventLogging()
Added a new function named TriggerEventLogging().  This function will take the arguments required for the logging specifications and neatly organze the parameters in the appropriate structure package.  This formatting will be required for when calling the TriggerEvent() function.  This function, TriggerEventLogging(), is treated as an accessor function as this will help organize the information without requiring other functions - outside from this function - to maintain with later changes during the evolution cycle.