Created a new class: Events
Created a new class named 'Events'.  This class contains all of the events so far that have been created within this software.  With that, all of the event functions that were in the class named IOCommon is now in the newly created class - - Events.  With that, the following functions have been shifted over:
 * RegisterEventLogging()
 * UnregisterEvent()
 * TriggerEvent()
 * TriggerEventLogging()

With this newly created class, this will help to organize all of the events in one class (in one file named Events.ps1).