Updated the project URL in the Help section
The Project URL was previously pointing at the Alphecca project, a Doom project.  Because the focus of PSCAT is not entirely going to be locked down 'one' specific project, it can be used publicly for various projects -- with such change in goals, it is best for the software to have its own repository.  With this software using its own repository, we can avoid version conflicts and centralize the code base - easy access to the latest revision (or commit).  Previously, PSCAT would carry over to the next in-house (G)ZDoom based project to another, this would cause conflicts as the versioning would be harder to track down and confusing code base due to conflicting versions.  Because I want this software to be used publicly, this software has to be far easier to find and fork-off into other projects.