Added new debug script: TraceStep
Created a new debug shellscript called TraceStep.ps1.  This shellscript will help setup the Set-PSDebug CMDLet for the current terminal session, which can either have PSDebug enabled with the requested flags or disabled entirely.  This will help to identify what is being executed within another program and what is taking place - at runtime.  This script was created to help me not forget about the CMDLet, I think this is a CMDLet that will be used again in the later future.