Thrashed superfluous conditional statement
I mistakenly added this conditional statement to make sure that the logging directories for the .NET Archive functionality existed, but there already existed a check for this already.  Thus, really, we where checking for the same directories through two different conditional statements -- we only needed one to do this operation efficiently.  I removed the latest conditional statement as we will not need this; only check if the directories exist rather than checking the ``$logging`` variable.