Updated SevenZip functions that interface with IOCommon
Updated all of the functions within the SevenZip object that utilizes the IOCommon object.  All functions that interfaces with the IOCommon within this class will require the ``$logging`` argument, which then will be used to call functions in IOCommon.

No longer call the function ``__CreateDirectories()`` in the class constructors as this will cause inconsistencies when the logging preferences differ.  If the program's logging is default to true - but the user's preference is set to false, one half of the operation is logged - the other is not.  Regardless if we enforce the logging as the default, we should still keep everything consistent.  Because we still need to assure that the ``__CreateDirectories()`` is used to create the necessary directories for the  Git Logs, we will instead invoke this function when wanting to perform any operation within the SevenZip object.  If ``$logging`` is ``true`` and ``__CreateDirectories()`` returns true - concurrently then the function performs as necessary, but if the ``__CreateDirectories()`` returns ``false`` - the function will abort when ``$logging`` is ``true``.