Provided documentation for error code -255 and -254
I have decided to add further documentation for what error codes will be returned when the function ``__ExecuteCommandRun()`` is called.  Because it is documented that a return code of ``-255`` is possible - but it is not actually provided in the ``ExecuteCommand()`` function, it is actually provided in the ``__ExecuteCommandRun()``.  Because of this possible confusion, this documentation should help others to understand where that '-255' will come from within the protocol.