Revised initialization for $fileOutput
At first, I wanted to use a ternary'ish way of initializing the variable named ``$fileOutput``, but the only way to pull it off is by using a traditional conditional statement - if\else.  With using the traditional if\else statement, it might seem a bit - - odd to see such a statement initialize the variable; even though it is perfectly legal and it does work - it just seems unfitting to the eyes and others may not like the way it is structured.  Instead, I have just revised the initialization to work within a conditional statement.