Revised WriteToFile()'s Out-File usage
Added two new property variables: ``$outFilePropertyWidth`` and ``$outFilePropertyEncoding``.
The variable: ``$outFilePropertyWidth`` handles the ``-Width`` argument in the Out-File CMDLet.  This allows so many characters per-line upon writing to the text file.
Additionally, the variable ``$outFilePropertyEncoding`` handles with the encoding property when using the Out-File CMDLet.  The encoding was previously already set to 'Default', nothing really changed other than setting the property having flexibility if needed in the future.