Resolved: Messages were forcefully logged
Resolved an issue where the messages that should be displayed to the user's screen, where forcefully logged regardless of the variable ``$logging`` state. When the state is now false, the messages from the function ``WriteToFile()`` will now display directly to the user's screen - instead of being logged.

NOTE: Depending on the scenario of the error, it is possible to reach a stackoverflow due to uncontrollable recursion.