Expunged superfluous variable $txtContent
Expunged a superfluous variable named ``$txtContent`` in the CreatePDFFile() function.  This variable was originally intended to hold all of the information in the Source File (the target text file), but it was far easier to just open the source file in the MS Word instance instead.  This variable essentially held space without ever being used in the function.