Provide error msg if repetition limit was hit
When attempting to determine a unique temporary directory name, it is possible (though highly unlikely) to have so duplicate directory names.  To help deter this from happening, a repetition counter is in place.  This repetition counter can help to avoid duplicate directory names, but there is a limit of how many directory can contain the same name - this program limits to fifty.  If the temp directory already contains 50 sub-directories of the exact same name, despite highly unlikely, the function will now provide a log report and notify the user.