Revised logging functionality in DeleteFile()
Instead of previously just throwing a vague message regarding if the operation was successful or not, provide a bit more information regarding the operation by state. If the task failed, provide the proper message and some information as to why it failed (directly from the PowerShell's pipe), likewise if the operation itself passed - provide a successful message. This will help to further investigate what could have caused the error while providing a cleaner and specific message.