Overhauled the logging functionality
Greatly revised the logging in the program; no longer will functions need the ``$logging`` parameter passed around like a crazed mad-man.  This will help to reduce the memory stack required to use the function.
Added a new locking mechanisim to further control what information is being logged.  This control is necessary to avoid possible recursive operations from occurring -- a stackoverflow will occur without such precautions.  However, the information that is logged IS NOT changed in this commit.