User can no longer configure Logging option
The user can no longer configure the option to enable or disable the logging functionality within the software, this is now configured by the program itself by the variable ``_DEBUGLOGGING_``.  This is a constant variable and will be the controlling factor for the logging functionality.

Added a new function in Logging named "DebugLoggingState()", which returns the value of the global variable ``_DEBUGLOGGING_``.

Removed the logging variable from the user's preferences.