Added a new function: WriteToLogFile()
Created a new static function named: WriteToLogFile().  This function will allow other objects to create their own logfiles (such as 7zip, Git, PS-Archive, etc) while also respecting the Logging availability and implementations.  Other objects, outside from Logging, shouldn't need to inspect the Logging object in order to log content to a specific file - infact other objects should only issue a request instead, this is where this function will come in to the picture.  This function will take the request but also check if the Logging functionality and availability is present at the current time upon the request, if the functionality is not available - the request is ignored, otherwise the request will be granted.  This approach is better in terms of maintenance and assuring that the objects can easily interact with other objects without the complexity.