Added debug information when extCMD ran sucessfully
Added debug information for when the external command was executed successfully.  However, this ignores the state of the extCMD.  Meaning, if the extCMD executed reached an error - it will still be treated as successfully.  What matters here is if Process.Start() actually executed the command successfully, regardless if the child extCMD reached a fault or not.