Provide debug info. when file doesn't in WriteToFile()
Provide debug information when the file could not be found after writting to it.  This will be useful to know if the file couldn't be found and why WriteToFile() is returning an error.

Learn something new!  It is not possible to return in a 'Finally' block, instead we will incorporate the $exitCode variable.  By default it will be '$true' to signify that everything is okay, but it will change if the file does not exist.