Non-populated descriptions are now sanitized!
It was possible to not provide any information in the description field when calling the ExecuteCommand() function, this would cause the ExecuteCommand() function to try to create one by combining the executable (or command) name and the arguments used, but doing this as-is can cause conflicts with the host's filesystem filename requirements.  The point of the description field is solely for the logging system, when a command is invoked - a new logfile is generated for the invoktion, with that the description is used as part of the filename: <<TimeStamp>>-<<Description>>.log.  Because of this, we must make sure that the description string does not contain any characters that could potentially cause a conflict with the host's filesystem.  To resolve any conflicts, a new function was added: __ExecuteCommandCreateDescription().  This function will remove any marked special characters from the description field and restrict the description to a specific length size.  By default, the description field may not exceed over sixty-four characters, any characters that exceed this limit will be discarded.