Resolved issue when displaying contents from an array
Resolved an issue when trying to present the contents from an array to the logfile.  This is due to assuming that using .ToString() would be sufficient - though only using the variable itself - with nothing more worked.  To assure that we would always provide the correct information, the array is properly evaluated by temporary storing the data from each index - into a temporary variable.  The temporary variable will be used to present the information to the logfile.