Provide additional information in DetectCommand()
Provide additional information regarding an application if it was detected by the host system.  The information gathered is provided by the output from Get-Command as an 'ApplicationInfo' object.  All information available will be from that object, but only some useful information is provided.  If the command was not found, then we use the original output of $null.