Resolved most of the line spacing inconsistencies
Resolved most of the inconsistencies when it comes to the spacing after the Additional Message within the logfile.  However, this only addresses most but not everything is entirely fixed.  Some of the automated strings that are generated are forced to carry the [CR][LF] regardless if their is incoming string.  Thus, if there is a list - most likely it will carry the extra [CR][LF] (or simply new-line).