Try to create destination dir. in CopyDirectory()
If the destination path does not exist, instead of ultmately giving up - try to create the path instead.  This will make this function a little bit easier to interface with when used by other objects\functions.  However, if the MakeDirectory() request fails to create the path, then the operation is aborted.