Revised the file's header comment-block
Revised the header comment-block to be a bit more clear and precise regarding the functionality within this object.  For example, it should much more clearer as to what is going on within this class and what dependencies this class requires in order to function properly.

Added notes for Dependencies and Developer Notes [API].  These should provide insight as to what the class requires or merely depends on - in order to function properly.