Changed 'object' to 'class' in __CreateDirectories()
Revised the wording used in the function's comment-block from 'object' to 'class'.  It seems like the same, but an object is an instance of the class and not an object at the code level.  Perhaps I am looking to deeply into this or perhaps it makes no pratical difference here.