Revised the VerifyArchive() function comment block
Greatly revised the comment block in the VerifyArchive() function.  This revision should greatly be transparent, descriptive, and elaborative regarding how this function operates.  The reason for this - is to assure that others understand (even for myself later on) that some limitations were faced during writing of the entire function.  Instead of just simply 'testing' an archive file, like with 7zip, the PowerShell module has it's own limits.  To accomplish the verification process, this function tests the archive file by extracting all of the data that is inside of the archive.