Revised ListFiles() header comment-block
Revised the ListFiles() function's comment-block that descripes the purpose of the function.  This should be a bit elaborative and easier to follow.
Added Dependencies to the comment-block.  It should be understood that we are using some .NET features in order to read the archive data file and also to inspect the data that is within the archive file.
Added Developer Notes to the comment block.  This addition contains remarks that we will be using the .NET framework instead of POSH CMDLets - as there is no CMDLet's available that will perform that task out-of-the-box.