Removed an unnessary comment; somehow missed this?
This should have been included in commit 1e9117452735d6556842ac087bf76f5f552fabc8 , but I missed it.  This exclusion is more in favor of better documentation - remove whatever that might be confusing or hard to decipher, in exchange for better documentation that is easier to follow and flows better with the code.