Fixed issue where $fileName was never set; changed some ini
Fixed an issue where the variable named $fileName was never initialized after commit 6cb67b0e244b0a038f332f8efa14ac8c35e85cf2
This was introduced as I was thinking of optimization later on, but my notes was lost.  Because of this, I completely forgot to reset this variable.
Because of $fileName, I decided to shift all of the initializations to another block of code -- $fileName was too long to put in the Declarations and Initialization.  With that, to make more sense - I have initialized the variables: $fileNameExt and $execReason within the same block as with $fileName.