Removed variable: $fileNameExt from ExtractArchive()
Removed the variable named: $fileNameExt.  This variable was used only once - and then forgotten about after used for $execReason.  Instead, the initialization - or the information that was previously held in $fileNameExt - is now immediately placed in $execReason.  Further, instead of using the POSH CMDLet Split-Path, we will just use the dotNET Core API to accomplish the task.  Consistency sakes as $fileName uses the dotNET Core API as well.