Removed two variables and replaced them with one variable
Expunged two variables named: $finalOutputPath and $cacheOutputPath.  These variables are no longer necessary in order to achieve our goal of generating a unique output directory for extracting the contents, we can - instead do this with just one variable instead.  I think at the time when I was constructing this function, I couldn't find an easy way to concatenate a string with an already existing - or didn't think about it at the time.  Either way, we can get a unique directory by concatenating the string as needed.