ExtractArchive() has cleaner output
When an archive data file has been successfully extracted, the logfile will have a better way of presenting the information to the reader.  Instead of throwing the list of files in the string, organize it by having a header, body, and footer.  This presentation is adjacent to the VerifyArchive() function.