Added two additional catches for ItemNotFound and FileFormat
Implemented two new additional catch's for the following exceptions: FileFormatException and ItemNotFoundException.  FileFormatException will be thrown when the data structure of the archive data file is malformed or does not contain the design as a traditional archive file.  ItemNotFoundException will be thrown when a specific data from the archive is corrupted.