[WIP] Revised the Declaration and Initializations
Greatly revised the declarations and initializations block of the code.  This revision comes with changes to the comments to each variable; the information given should be a bit easier to understand.
Removed two variables: $cacheArchiveFileName && $finalArchiveFileName, but replaced with one variable: $archiveFileName.  This change, despite presently breaking this function, will help to optimize resources required to generate the archive file's name.
Relocated the initialization of the variables: $targetDirectoryFiltered and $execReason below the initialization and Declarations block as the initializing them requires a lot of space (horizontally) and can be a bit hard to read and follow; this change should help.