execSTDOUT only contains one output file
Only one file will exist in the output variable for 'execSTDOUT', this is because we are only creating just 'one' archive datafile when using the CreateArchive() function.  With this in mind, we will not need to use a for-loop to iterate through an array when the array only contains just one element.
In addition, the output message is a bit nicer to read.  Instead of just "File: " it will now show "Newly created archive datafile path: ".