Expanded CheckPathExists() functionality to non-literal paths
Expanded the functionality of CheckPathExists() in the IOCommon object to accept non-literal paths.  This change, despite effecting all other parts of the program - coupled a bit, the functionality previously does not change in this revision.  This expanded function now allows the possibilities to use wildcards, but also relative paths - despite not recommended to use relative paths.