CreateArchive() allows wildcards for leaf files
When wanting to create a new archive datafile, if the target directory path contains a wildcard - this function will now support it freely.  This can be useful if limiting to just a specific file range or an open range for all files within the target path.

Because of this change, '$targetDirectoryFiltered' is has been expunged.  We will now work directly with the $targetDirectory variable.