MakeDirectory() will now report $true if directory already exists
The function named, MakeDirectory() in the IOCommon object, will now return $true if the directory already exists instead of returning $false.  $false should only be provided when the directory could not be created due to /whatever/ reason, it should not provide $false when the directory already exists.  Furthermore, because the documentation hints that this functionality was supposed to be already supported, I believe it might have been accidentally revised otherwise.  Due to limited bandwidth, I am unable to determine where this issue first arised.