Revised the logic in ThrashLogs()
Revised the logic when trying to expunge the reports and\or just the log files.  The previous logic, when wanting to remove both the log and report files, if an error was reached during after attempting to delete the log files - the function immediately stops.  Now, the function will try to delete the log and report files, but will return a failure if one of the operations (or both) failed.  This assures that the operation was attempted.