Replaced __SupportedHashAlgorithms() for an Enumorator
Removed the function named __SuppotedHashAlgorithms() from the 7Zip object.  This function's purpose can be easily fulfilled by using an enumator instead - of a function that scans through a list.  Despite the list being extremely short and requires literally no hinderance in performance, if the list were to grow over time - performance could be a factor in the future.  With this change, ArchiveFile() required some modifications.
 - Changed the argument for $hashAlgorithm from a string to the enumerator 'FileHashAlgorithm7Zip'.
 - Removed the required call to __SupportedHashAlgorithms() as this function no longer exists.

Added the enumerator for known file hash algorithms.