FileList() now provides logfile
FileList() from the Default Compression (.NET) object, will now try to create a logfile regarding the operation.  Thus, the list of files will now be present within the program's logfile as well as the object logfile.  Despite no real CMDLet is used, it is still appropriate to provide the consistency through out the program.