ListFiles() now reports when output is NULL
After executing 7Zip, the output was checked to assure that something meaningful was captured.  If incase the output was empty (or null in this cause), the event will now be logged as a 'Warning'.  Though, if this issue comes up in a real life scenario, please report this as a bug.  I doubt it'll be an issue, but report it none the less.