[BROKEN] Revised the variables
CreateArchive() is presently broken and may cause the program to fail as a whole.  Please give me time to update this function piece by piece.

Renamed $finalArchiveFileName to $archiveFileName.  It is rather unnessary to attach 'final' as part of the variable name.  This was previously used in conjunction with the $cacheArchiveFileName.

Commented out the variable $cacheArchiveFileName.  I believe that it is superfluous it is possible to omit its purposes.  We can, however, just use $archiveFileName directly instead.

Revised the variables documentation to be much more readable.  I was not really happy with some of the descriptions that was previously provided.