Updated the Class Description
Revised the Class Description to be a bit more thorough regarding what activities take place within this class, as well as pointing out - it is optional to utilizes if the external software is not detected.  The dependencies provided are only popular software, but it is not limited to just those listed.