Added a new function: FindGit()
Added a new function called: FindGit().  This function, although not complete, will try to find the Git executable automatically for the End-User.  To acomplish this task, there will be a list of common paths of where the executable may reside - each path will be inspected.  If the Git executable can be found - the first successful hit will be the that is returned to the calling function.

TODO:
  Add more paths into the list, this requires some investigating.