FindGit() will now properly scans and returns a positive hit
The FindGit() function will now be able to properly scan the search results and return the first positive hit back to the calling function as intended.  If no results were found, then $null will be returned instead.

This is still a WIP.  It is taking a lot longer than usual to get things moving....