Corrected the return descriptions used
Renamed 'Status' to 'Operation Status'.
Corrected the documentation for false and true return codes, the return codes has nothing todo with the executable being found or not.  Despite it is required to use the executable, the main heart of the function is to update the localized repository - not finding the executable binary file.