Work around with the parameter issue
Essentially undid what I did in the previous commits regarding the negative\parameter denotion.  When providing the $changelogSize from the user's settings - apply the parameter denotion.  If the the setting is already negated - then correct it through the cache variable so that it may be used in the log output.  The cache variable containing the correction - will then be applied to the $changelogSize variable.