If the _DEBUGLOGGING_ is false, do not log the internal message.
If the _DEBUGLOGGING_ is set as false within the current program's environment, then the program's logging functionality will be disabled.  However, this only reflects torwards the internal program's logging as well as External CMD logging.  To disable the essential logging witihin this program - flip this variable '_DEBUGLOGGING_' to false.  Though, for debug purposes and how dynamic environments can be across different systems - it is recommended to leave this variable true.  Only change this variable if you are certain you will never run across a problem.