Do not check the logging state in this manner anymore.
Instead of checking the _DEBUGLOGGING_ from functions that are not related to the Logging object,  have only the _DEBUGLOGGING_ checked by the Logging object only.  This way outside functions can interact with the functions that are most important -- only the internal functions can handle with the other necessities as needed.