Removed superfluous objects in Load()
Removed the User Preferences object, Git Control Object, 7Zip Object, and the Default Compress Object.  These were possibly going to be used in such a way that the objects would be returned after manipulation, but we can achieve the same technique using references (pointers) of the objects provided in the function arguments.