Added a compatibility check
This compatibility check will make sure that the PowerShell host can be able to run the script correctly.  The check will only consider: Core and Legacy.  If the end-user is using the PowerShell Core, then 'Core' is provided - that host is capable of running the script.  However, if one is using Windows PowerShell, then 'Legacy' is provided.  Once 'Legacy' had been detected, the script will throw an error and stop the program from running any further.

This check may not be fully reached with legacy shell versions, but if and only if it helps deter versions that managed to reach this point.