Revised main()'s function documentation
Compatibility checks are now done outside from main().  If incase the shell manages to read all the way to the main function call, terminate it before the OOP aspect of the program begins.  This way the shell will safely terminate before reaching a critical error - unless it already reached a fault before hand.