Do not use 'return' to exit from the program, but use 'exit' instead.
Do not use the 'return' statement to close the application, which merely throws the exit code onto the terminal - instead of the POSH's environment.  Instead, use the 'exit' statement to exit the program and to properly issue the exit code to the POSH's environment.

The Exit Code for Compatibility Issue Determined may change in the future once the program had been further developed.