Added two new functions related to the Window Title
Added two new functions: GetTerminalWindowTitle() and SetTerminalWindowTitle().  GetTerminalWindowTitle() allows the ability to retrieve the Window's current title, Where as the SetTerminalWindowTitle() sets the window's title as specified with the first argument.

Added a new region named "Terminal Window Control".  This will help organize functions that provides the ability to alter or retrieve settings regarding the window's environment in which the program is being executed.