Added a comment regarding the various ways to end the program
The user is not just restricted to 'X' to exit the application, they can also type in "Exit" or "Quit".  These options can be valuable if the user is in a panic and needs to leave the program - because they are freaking out over the new environment (Not a lot of users will utilize programs within the CUI environment).  But there could also be advanced users that are used to typing 'Exit' or 'Quit' as well.  This option should add a slight benefit for such rare occasions.