Added the ability to access generalized websites
Added the ability to access websites while respecting the user's setting.  While this function is still WIP, the main principle of the function remains the same.  Further, the main function that handles the operation is emulated as 'Private', there are two public functions that will act as a bridge.  The reason for this change is so that we can provide what details are necessary for the operation, nothing more as it will be confusing to the user side of the program.