EvaluateExecuteUserRequest() now requires objects passed by pointers
EvaluateExecuteUserRequest() will now requires the following arguments to be passed as pointers (references):
 - $userPreferences
 - $gitControl
 - $sevenZip
 - $defaultCompress

This is required as some options within the main menu will allow the ability to reconfigure the objects, such as the Preferences option.  However, not every option requires full write privileges with the objects.  If wanted to view a value or access a set of data with mere read privileges only, all that is needed is to pass by value to the options.