Removing Project Version from Access Web Site
More that I think about it, the only time when it is necessary to perform an update - is when we want to update the program itself, the project will be updated via source control.  There's no real need to centralize the update functionality in a way to update anything and everything, just focus on the program.