Revised the sectional border algorithm again
Second thought, lets do this!  The algorithm is really simple:  5 + Title / 2 + 5.  Meaning, that before the title, there should be 5 characters of the border displayed.  Then the length of the title divided by two - is the length of the character border length.  After the title, then five final border characters.

I love it!