Forgot to apply a 'Catch' condition in DetectCommand()
When using the Get-Command with the STOP flag used in -ErrorAction, this will cause the entire program to halt and terminate with an error.  But that is only true, if there was no proper Catch condition.  Now, the Catch added into the algorithm, Get-Command will not cause the program to halt when it returns an error.